merge r4459 from branch 2.0 to trunk
feature 1316 added: we now have the ability to only fill the single value
properties that are currently empty in the database instead of automatically
replacing.

feature 1312: as a matter of consistency, I have changed the replace_mode
(that was only for multiple values properties) into multiple_value_mode +
single_value_mode.