feature 1507: ability to install themes directly from piwigo.org/ext.
The theme manager backend for installation is the plugin manager backend,
because they are very close.

thanks to feature:1502, installing a theme and installing a plugin are very
similare operations. The only differences are the installation directory and
the main file to find in the archive.

This is only the very first step, the "add new theme" screen needs many
improvements, to begin with display style.