merge r6320 from trunk to branch 2.1
bug 1667 fixed: r6027 was fixing a minor error on tabs for Google Chrome/Safari
BUT was introducing a major visual issue for IE8. Backmerged + adds the -webkit
radius anyway.

This is a quickfix for release 2.1.1, the improved display introduced by nikrou
was nice and it would be better to have it, but working in IE8