Tag for 0.14 version of jiwigo This version integrates all jiwigo-ws-api changes : - Dependency injection (jiwigo now uses spring for injection) - Better exception management (thanks to anthony https://github.com/anthonydahanne/) - Better support for android (thanks to anthony too) - Suppression of jdom
The images are not displayed in a new frame anymore
It also integrates a new way of sending images : a panel that displays the user's file system (filters image files). The next step will be to send images in background
The comments dialog has been changed to be more "beautiful".
The connection via a proxy is now possible